Change structure of YAML to add the future flow
Set the current values in the YAML to be the default value.

Add values for the future benefit cap to the YAML..

Change the structure of the YAML file. Describe the current values as the
"default" values and add the future configuration to the YAML file

Add the data for the benefit caps by region and the values for exempt and
non-exempt benefits to a "future" node in the YAML file.