Simplify Services.worldwide_api method
There is already a mechanism available for overriding the URI for specific
services, so there is no need to have some special code to do this. Service
URI's are automatically looked up via `Plek` and they can be overridden using
environment variables of the form: `PLEK_SERVICE_<service-name>_URI` (see [1]).

In this case, we can set the `PLEK_SERVICE_WHITEHALL_ADMIN_URI` environment
variable to "https://www.gov.uk" if we want to use the production website to
supply the data for the Worldwide API.

Although I've updated the `developing-without-vm` documentation, I haven't
yet updated the `developing-using-vm` documentation, because I don't use the VM
myself and so am not sure how the documentation should be changed. I have,
however, confirmed that it is possible to use a locally-running version of
the `whitehall` app to serve the Worldwide API data.

I have also updated the `common-errors` documentation to reflect the error that
occurs if the Worldwide API is not pointing at an appropriate URI.

[1]: https://github.com/alphagov/plek/blob/b1f917cd6498aea8867728dab0132d0373ca702b/README.md#for-base-urls