Fix income limit - use limit for current tax year
The original `is_before_april_changes` calculate block was removed in this
commit [1], but then re-introduced in this commit [2] because the state variable
was still being used in the `whats_the_husbands_income?` and
`whats_the_highest_earners_income?` questions.

I suspect that the ternary condition used in these two questions to choose
an income limit from two hard-coded values should've been replaced in the
earlier of these two commits by a lookup of the
`income_limit_for_personal_allowances` value from the `DataQuery` introduced
in that commit.

I think this means that `calculate-married-couples-allowance` will have been
always using an income limit of £27,000 in the `next_node` blocks since this
commit. This value is the income limit for the 2014-2015 tax year [3]. Since
then the income limit has risen to £27,700.

This commit fixes the problem so that the correct income limit is used for the
current tax year.

Note that this is a relatively subtle bug, because
`MarriedCouplesAllowanceCalculator#calculate_allowance` has been using the
*correct* value since the earlier commit.

[1]: 59cd29fb51435bb53db2efa9a8a289f44185b380
[2]: 691c024213c8e673cc8179ecb218fc5991bf0777